+++ internal/server.go
@@ -31,7 +31,7 @@ func (s *Server) Start() {
 	httpAddress := fmt.Sprintf(":%d", s.config.HttpPort)
 	httpsAddress := fmt.Sprintf(":%d", s.config.HttpsPort)
 
-	if s.config.SSLDomain != "" {
+	if s.config.TLSDomain != "" {
 		manager := s.certManager()
 
 		s.httpServer = s.defaultHttpServer(httpAddress)
@@ -44,7 +44,7 @@ func (s *Server) Start() {
 		go s.httpServer.ListenAndServe()
 		go s.httpsServer.ListenAndServeTLS("", "")
 
-		slog.Info("Server started", "http", httpAddress, "https", httpsAddress, "ssl_domain", s.config.SSLDomain)
+		slog.Info("Server started", "http", httpAddress, "https", httpsAddress, "tls_domain", s.config.TLSDomain)
 	} else {
 		s.httpsServer = nil
 		s.httpServer = s.defaultHttpServer(httpAddress)
@@ -73,13 +73,13 @@ func (s *Server) certManager() *autocert.Manager {
 	client := &acme.Client{DirectoryURL: s.config.ACMEDirectoryURL}
 	binding := s.externalAccountBinding()
 
-	slog.Debug("SSL: initializing", "directory", client.DirectoryURL, "using_eab", binding != nil)
+	slog.Debug("TLS: initializing", "directory", client.DirectoryURL, "using_eab", binding != nil)
 
 	return &autocert.Manager{
 		Cache:                  autocert.DirCache(s.config.StoragePath),
 		Client:                 client,
 		ExternalAccountBinding: binding,
-		HostPolicy:             autocert.HostWhitelist(s.config.SSLDomain),
+		HostPolicy:             autocert.HostWhitelist(s.config.TLSDomain),
 		Prompt:                 autocert.AcceptTOS,
 	}
 }